can create hcmio, update mstmb.

+++ esmp_demo/src/main/java/com/practice/esmp_demo/model/entity/Tcnud.java
@@ -26,15 +26,15 @@ public class Tcnud {
     @Column(name = "Stock")
     private String stock;
     @Column(name = "Price")
-    private int price;
+    private double price;
     @Column(name = "Qty")
-    private int qty;
+    private double qty;
     @Column(name = "RemainQty")
-    private int remainQty;
+    private double remainQty;
     @Column(name = "Fee")
-    private int fee;
+    private double fee;
     @Column(name = "Cost")
-    private int cost;
+    private double cost;
     @Column(name = "ModDate")
     private String modDate;
     @Column(name = "ModTime")
